Simplify pricing architecture: two-layer model, delete metric: cost
Collapses four overlapping pricing config entry points (top-level
pricing.rates/aliases, external pricing.yaml supplement, provider
map/overrides, standard/curated tables) into one: providers[].pricing
{currency, aliases, rates}, resolved against the built-in standard table.
Top-level pricing: block is gone entirely, replaced by a top-level
exchange_rate: map with a built-in default table as fallback
(internal/pricing/standard_exchange_rate.yaml).

Deletes metric: cost and everything that existed only to support it:
core.Endpoint.PricingRate, pricing.FoldSpec, quota.Counters.Cost/ChargeCost,
config.Config.ResolvedPricing/PricingAccounting. The routing/quota hot path
now carries zero price data and zero currency concepts — pricing resolution
lives entirely in the offline vmr report/vmr analyze path
(internal/pricing.Resolver, unchanged API). A Credits/money-style budget is
now expressed as a tokens Limit (convert once: budget / price), with
model_multipliers/token_weights carrying the per-model/per-component
differentiation metric: cost used to provide.

Breaking change. See
docs/future-strategy/pricing_architecture_simplification_plan.md for the
full rationale and docs/future-strategy/pricing_architecture_action_plan.md
for the implementation tracker.

+++ internal/report/findings_quota_test.go
@@ -37,7 +37,7 @@ func TestQuotaExhaustionFinding_BelowThresholdDoesNotFire(t *testing.T) {
 // otherwise fabricate an alert — an estimate must never be the basis of one.
 func TestQuotaExhaustionFinding_NoLiveDataDoesNotFire(t *testing.T) {
 	rep := &Report2{ProviderQuotas: []ProviderQuotaRow{
-		{Provider: "acct1", WindowConsumed: f64(999999), Live: nil},
+		{Provider: "acct1", WindowConsumed: 999999, Live: nil},
 	}}
 	if f := quotaExhaustionFinding(rep, i18n.EN); f != nil {
 		t.Fatalf("expected no finding when Live is nil, regardless of WindowConsumed, got %+v", f)
